# attached_assets/openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Slack Clone API",
      "version": "v1.0.1",
      "description": "Slack-like messaging platform with universal soft-deletes, selective partitioning, email-based verification, triggers for updatedAt, real-time features, error code enumeration, and recommended monorepo structure."
    },
    "servers": [
      {
        "url": "https://api.your-slack-clone.com/v1",
        "description": "Production server (v1)"
      },
      {
        "url": "http://localhost:3000/v1",
        "description": "Local development server (v1)"
      }
    ],
...
    "paths": {
...
      "/emojis": {
        "get": {
          "tags": ["Emojis"],
          "summary": "List all emojis (by default excludes deleted unless specified)",
          "parameters": [
            {
              "name": "includeDeleted",
              "in": "query",
              "schema": { "type": "boolean", "default": false }
            }
          ],
          "responses": {
            "200": {
              "description": "Array of emojis",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": { "$ref": "#/components/schemas/Emoji" }
                  }
                }
              }
            }
          }
        },
        "post": {
          "tags": ["Emojis"],
          "summary": "Add a new custom emoji",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EmojiCreateRequest" }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Emoji created",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/Emoji" }
                }
              }
            }
          }
        }
      },
      "/emojis/{emojiId}": {
        "get": {
          "tags": ["Emojis"],
          "summary": "Get specific emoji details",
          "parameters": [
            {
              "name": "emojiId",
              "in": "path",
              "required": true,
              "schema": { "type": "integer" }
            }
          ],
          "responses": {
            "200": {
              "description": "Emoji details",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/Emoji" }
                }
              },
              "404": {
                "description": "EMOJI_NOT_FOUND",
                "content": {
                  "application/json": {
                    "schema": { "$ref": "#/components/schemas/ErrorResponse" }
                  }
                }
              }
            }
          }
        },
        "delete": {
          "tags": ["Emojis"],
          "summary": "Soft-delete an emoji",
          "description": "Sets deleted=true instead of physical removal.",
          "parameters": [
            {
              "name": "emojiId",
              "in": "path",
              "required": true,
              "schema": { "type": "integer" }
            }
          ],
          "responses": {
            "204": { "description": "Emoji soft-deleted" },
            "404": {
              "description": "EMOJI_NOT_FOUND",
              "content": {
                "application/json": {
                  "schema": { "$ref": "#/components/schemas/ErrorResponse" }
                }
              }
            }
          }
        }
      },
...
    },
    "components": {
      "securitySchemes": {
        "BearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      },
      "schemas": {
...
        "ErrorResponse": {
          "type": "object",
          "properties": {
            "error": { "type": "string" },
            "details": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "enum": [
                    "USER_NOT_FOUND",
                    "WORKSPACE_NOT_FOUND",
                    "CHANNEL_NOT_FOUND",
                    "MESSAGE_NOT_FOUND",
                    "EMOJI_NOT_FOUND",
                    "INVALID_CREDENTIALS",
                    "UNAUTHORIZED",
                    "DUPLICATE_REACTION",
                    "FORBIDDEN"
                  ]
                },
                "message": { "type": "string" }
              }
            }
          }
        },
...
        "Emoji": {
          "type": "object",
          "properties": {
            "emojiId": { "type": "integer" },
            "code": { "type": "string" },
            "deleted": { "type": "boolean" },
            "createdAt": { "type": "string", "format": "date-time" },
            "updatedAt": { "type": "string", "format": "date-time" }
          }
        },
        "EmojiCreateRequest": {
          "type": "object",
          "properties": {
            "code": { "type": "string" }
          },
          "required": ["code"]
        },
...
      }
    }
  }